* origin/released:
  pixman xkbcomp libX11 libXau mesa fontconfig xserver xkeyboard-config git update 7 jan 2013

Conflicts:
	fontconfig/src/fcatomic.c
	pixman/pixman/pixman-mmx.c
	pixman/pixman/pixman-sse2.c
	xorg-server/dix/dispatch.c
	xorg-server/hw/xwin/wincursor.c
	xorg-server/hw/xwin/winmsg.c
	xorg-server/hw/xwin/winscrinit.c
	xorg-server/hw/xwin/winsetsp.c
	xorg-server/hw/xwin/winwin32rootless.c
	xorg-server/xfixes/cursor.c

+++ libXau/AuFileName.c
@@ -45,12 +45,12 @@ XauFileName (void)
 {
     const char *slashDotXauthority = "/.Xauthority";
     char    *name;
-    static int	bsize;
+    static size_t	bsize;
     static int atexit_registered = 0;
 #ifdef WIN32
     char    dir[128];
 #endif
-    int	    size;
+    size_t  size;
 
     if ((name = getenv ("XAUTHORITY")))
 	return name;
@@ -70,7 +70,7 @@ XauFileName (void)
     if (size > bsize) {
 	if (buf)
 	    free (buf);
-	buf = malloc ((unsigned) size);
+	buf = malloc (size);
 	if (!buf)
 	    return NULL;
